kf6-cmake: fix several "contains reference to TMPDIR" warnings
Introduce postprocessing of cpp/h files that are generated during
compilation and print absolute source code paths as commetns. This
replacement uses the heursitic that such comments always have a space in
front of the path (unlike #include or macros).
If wrong paths are catched this might need an extended regexp logic to
ensure that only strings inside C++ comments are replaced.