vlc: fix build failure when wayland-scanner is missing from host
Our internal CI system failed on vlc because of a missing binary.

Turns out the build system was looking for 'wayland-scanner' on the
host instead of the sysroot-native directory.

This patch changes the default detection mecanism and instead provides
a way to set the WAYLAND_SCANNER variable from the bitbake recipe.

Signed-off-by: Marc Ferland <marc.ferland@sonatest.com>